+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Added a global variable that stores `rowInfo` as well as sorting & highlighting info, using `React.Context`.
 - Added a mechanism to sort rows of a HiGrass `horizontal-multivec` track and vertical tracks on the left and right sides which is activated when users click on sorting buttons.
 - Added a feature to interactively highlight rows by typing keywords on a text field which appears when a search button is clicked.
+- Added a feature to interactively resize the width of each vertical track by dragging resizers.
 
 ### Changed
 - Moved to rendering the row info categorical metadata on a single `<canvas/>` element, rather than colored HTML elements for each row.